inforadar_parser/requirements.txt
requests
pymysql
pysocks
playwright
pymysql
cryptography
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.33
